Solution to Check Even or Odd

+++ Check Even or Odd@@6788dd63490637cfc44e1b8d/code.c
@@ -2,9 +2,9 @@
 #include <stdio.h>
 int main()
 {
-    float a;
-    scanf("%f",&a);
-    if (a%2==0)
+    int a;
+    scanf("%d",&a);
+    if (a % 2 == 0)
     printf("Even");
     else printf("Odd");
     return 0;
